Refactor: Register custom hook
